refactor: inline MySQL CLI unit tests
