--- src/infra/adapters/mysql/cli/error_tests.rs
@@ -1,54 +1,0 @@
-#[cfg(test)]
-mod error_tests {
-    use super::super::*;
-    use sabiql_app::model::connection::error::{ConnectionErrorInfo, ConnectionErrorKind};
-
-    #[test]
-    fn classifies_mysql_query_failures_by_server_error() {
-        assert!(matches!(
-            classify_mysql_query_failure(
-                b"ERROR 1045 (28000): Access denied for user 'app'@'localhost' (using password: YES)"
-            ),
-            DbOperationError::ConnectionFailed(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(b"ERROR 1142 (42000): command denied to user"),
-            DbOperationError::PermissionDenied(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(
-                b"ERROR 1044 (42000): Access denied for user 'app' to database 'app'"
-            ),
-            DbOperationError::PermissionDenied(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(b"ERROR 1146 (42S02): Table does not exist"),
-            DbOperationError::ObjectMissing(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(b"ERROR 1205 (HY000): Lock wait timeout exceeded"),
-            DbOperationError::LockTimeout(_)
-        ));
-        assert!(matches!(
-            classify_mysql_query_failure(
-                b"ERROR 1452 (23000): Cannot add or update a child row: a foreign key constraint fails"
-            ),
-            DbOperationError::ForeignKeyViolation(_)
-        ));
-        let masked = classify_mysql_query_failure(b"ERROR password=secret");
-        assert!(!masked.masked_details().contains("secret"));
-    }
-
-    #[test]
-    fn classifies_mysql_tls_query_failures_as_connection_errors() {
-        let error = classify_mysql_query_failure(
-            b"ERROR 2026 (HY000): SSL connection error: error:0A000086:SSL routines::certificate verify failed",
-        );
-
-        assert_eq!(
-            ConnectionErrorInfo::from_db_operation_error(&error).kind,
-            ConnectionErrorKind::MySqlTlsHandshakeFailed
-        );
-        assert!(matches!(error, DbOperationError::ConnectionFailed(_)));
-    }
-}
